+++ frontend/src/types/dictionaryPlant.ts
@@ -20,3 +20,8 @@ export type DictionaryPlantNameSearchResult = Pick<DictionaryPlant, 'id' | 'name
 export type Season = 'spring' | 'summer' | 'autumn' | 'winter';
 export type SeasonKor = '봄' | '여름' | '가을' | '겨울';
 export type ManageLevel = '초보자' | '경험자' | '전문가' | '정보없음';
+
+export interface DictionaryPlantRegistrationForm {
+  name?: string;
+  image?: File;
+}
