--- frontend/src/types/dictionaryPlantRegistration.ts
@@ -1,4 +1,0 @@
-export interface DictionaryPlantRegistrationForm {
-  name?: string;
-  image?: File;
-}
