--- frontend/src/types/image.ts
@@ -1,4 +1,0 @@
-export interface ImageFormData<T> {
-  requestForm: T;
-  imageData: Blob | null;
-}
